# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API Sistem Informasi Akademik Tanya UNPRA
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List mata kuliah (semua role)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: prodi
          in: query
          schema: { type: string }
        - name: semester
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Daftar mata kuliah
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoursesResponse"
        "401":
          description: Tidak terautentikasi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Tambah mata kuliah baru (admin & dosen)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
      responses:
        "201":
          description: Mata kuliah berhasil dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseResponse"
        "400":
          description: Data tidak valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Akses ditolak
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Kode mata kuliah sudah digunakan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Detail mata kuliah (semua role)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Detail mata kuliah
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseResponse"
        "404":
          description: Mata kuliah tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update mata kuliah (admin & dosen)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseRequest"
      responses:
        "200":
          description: Mata kuliah berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseResponse"
        "404":
          description: Mata kuliah tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Hapus mata kuliah (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Mata kuliah berhasil dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Mata kuliah tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /lecturers:
    get:
      operationId: listLecturers
      tags: [lecturers]
      summary: List dosen (semua role)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: prodi
          in: query
          schema: { type: string }
        - name: fakultas
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Daftar dosen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LecturersResponse"
  /lecturers/{id}:
    get:
      operationId: getLecturer
      tags: [lecturers]
      summary: Detail data dosen (semua role)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Detail dosen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LecturerResponse"
        "404":
          description: Dosen tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateLecturer
      tags: [lecturers]
      summary: Update data dosen (dosen edit sendiri, admin edit semua)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLecturerRequest"
      responses:
        "200":
          description: Data dosen berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LecturerResponse"
        "403":
          description: Dosen hanya bisa edit profil sendiri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Dosen tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLecturer
      tags: [lecturers]
      summary: Hapus data dosen (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Data dosen berhasil dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Dosen tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]
    Course:
      type: object
      properties:
        id:
          type: string
        kode:
          type: string
        nama:
          type: string
        sks:
          type: integer
        semester:
          type: integer
        prodi:
          type: string
        deskripsi:
          type: string
          nullable: true
        lecturerId:
          type: string
          nullable: true
        lecturerName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, kode, nama, sks, semester, prodi, createdAt, updatedAt]
    CourseResponse:
      type: object
      properties:
        course:
          $ref: "#/components/schemas/Course"
      required: [course]
    CoursesResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [courses, pagination]
    CreateCourseRequest:
      type: object
      properties:
        kode:
          type: string
          minLength: 2
          maxLength: 20
        nama:
          type: string
          minLength: 3
          maxLength: 200
        sks:
          type: integer
          minimum: 1
          maximum: 6
        semester:
          type: integer
          minimum: 1
          maximum: 14
        prodi:
          type: string
          minLength: 2
          maxLength: 100
        deskripsi:
          type: string
          maxLength: 1000
        lecturerId:
          type: string
          format: uuid
      required: [kode, nama, sks, semester, prodi]
    UpdateCourseRequest:
      type: object
      properties:
        kode:
          type: string
        nama:
          type: string
        sks:
          type: integer
        semester:
          type: integer
        prodi:
          type: string
        deskripsi:
          type: string
        lecturerId:
          type: string
          format: uuid
...
    Lecturer:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        nidn:
          type: string
        prodi:
          type: string
        fakultas:
          type: string
        jabatan:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        expertise:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, nidn, prodi, fakultas, createdAt, updatedAt]
    LecturerResponse:
      type: object
      properties:
        lecturer:
          $ref: "#/components/schemas/Lecturer"
      required: [lecturer]
    LecturersResponse:
      type: object
      properties:
        lecturers:
          type: array
          items:
            $ref: "#/components/schemas/Lecturer"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [lecturers, pagination]
    UpdateLecturerRequest:
      type: object
      properties:
        prodi:
          type: string
        fakultas:
          type: string
        jabatan:
          type: string
        phone:
          type: string
        expertise:
          type: string
...